Modify handleLoadFailure function
