+++ src/Models/IoTWorkspaceProject.ts
@@ -467,31 +467,4 @@ export class IoTWorkspaceProject extends IoTWorkbenchProjectBase {
       throw error;
     }
   }
-
-  /**
-   * If current folder is an IoT Workspace Project but not open correctly, ask
-   * and open the IoT Workspace Project. Return true. otherwise return false.
-   */
-  static async handleIoTWorkspaceProjectFolder(
-      telemetryContext: TelemetryContext): Promise<boolean> {
-    if (!vscode.workspace.workspaceFolders ||
-        !vscode.workspace.workspaceFolders[0]) {
-      return false;
-    }
-
-    const rootPath = vscode.workspace.workspaceFolders[0].uri.fsPath;
-    const workbenchFileName =
-        path.join(rootPath, 'Device', FileNames.iotworkbenchprojectFileName);
-
-    const workspaceFiles = fs.readdirSync(rootPath).filter(
-        file => path.extname(file).endsWith(FileNames.workspaceExtensionName));
-
-    if (fs.existsSync(workbenchFileName) && workspaceFiles &&
-        workspaceFiles[0]) {
-      await askAndOpenProject(rootPath, workspaceFiles[0], telemetryContext);
-      return true;
-    }
-
-    return false;
-  }
 }
